requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
mysqlclient==2.2.0
django-filter==23.5
PyJWT==2.8.0
bcrypt==4.1.2
